requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
fbs==0.9.0
Pillow==8.0.1
PyQt5==5.9.2
requests==2.25.1